src/main/java/com/compass/javaapisamp/controller/validator/DateStringValidator.java
package com.compass.javaapisamp.controller.validator;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;
import java.time.LocalDate;
import java.time.format.DateTimeFormatter;

public class DateStringValidator implements ConstraintValidator<DateString, String> {

    private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyyMMdd");

    @Override
    public boolean isValid(String value, ConstraintValidatorContext context) {
        try {
            LocalDate.parse(value, formatter);
        }catch (Exception e) {
            return false;
        }
        return true;
    }
}
